[PATCH] direct-IO return type fixes
From: me, Badari Pulavarty <pbadari@us.ibm.com>

Currently a direct-IO read or write of more than 2G on 64-bit machines is
broken.  Replace int with ssize_t in various places to fix that up.

+++ fs/hfsplus/inode.c
@@ -114,8 +114,8 @@ static int hfsplus_get_blocks(struct inode *inode, sector_t iblock, unsigned lon
 	return ret;
 }
 
-static int hfsplus_direct_IO(int rw, struct kiocb *iocb, const struct iovec *iov,
-			  loff_t offset, unsigned long nr_segs)
+static ssize_t hfsplus_direct_IO(int rw, struct kiocb *iocb,
+		const struct iovec *iov, loff_t offset, unsigned long nr_segs)
 {
 	struct file *file = iocb->ki_filp;
 	struct inode *inode = file->f_dentry->d_inode->i_mapping->host;
